Add async interceptors for Insert/Update/Delete returning and fix deleteReturning operations
Agent-Logs-Url: https://github.com/micronaut-projects/micronaut-data/sessions/ef3621ec-dc08-44a5-8018-e3d6416502d6

+++ data-runtime/src/main/java/io/micronaut/data/runtime/intercept/reactive/DefaultDeleteReturningManyReactiveInterceptor.java
@@ -41,15 +41,20 @@ public final class DefaultDeleteReturningManyReactiveInterceptor extends Abstrac
 
     @Override
     protected Publisher<?> interceptPublisher(RepositoryMethodKey methodKey, MethodInvocationContext<Object, Object> context) {
-        Optional<Object> deleteEntity = findEntityParameter(context, Object.class);
-        if (deleteEntity.isPresent()) {
-            return reactiveOperations.deleteReturning(getDeleteReturningOperation(context, deleteEntity.get()));
-        }
-        Optional<Iterable<Object>> deleteEntities = findEntitiesParameter(context, Object.class);
-        if (deleteEntities.isPresent()) {
-            return reactiveOperations.deleteAllReturning(getDeleteReturningBatchOperation(context, deleteEntities.get()));
-        }
         PreparedQuery<?, Object> preparedQuery = prepareQuery(methodKey, context);
+        Class<?> rootEntityType = preparedQuery.getRootEntity();
+        Class<?> resultType = preparedQuery.getResultType();
+
+        if (resultType == rootEntityType) {
+            Optional<Object> deleteEntity = findEntityParameter(context, Object.class);
+            if (deleteEntity.isPresent()) {
+                return reactiveOperations.deleteReturning(getDeleteReturningOperation(context, deleteEntity.get()));
+            }
+            Optional<Iterable<Object>> deleteEntities = findEntitiesParameter(context, Object.class);
+            if (deleteEntities.isPresent()) {
+                return reactiveOperations.deleteAllReturning(getDeleteReturningBatchOperation(context, deleteEntities.get()));
+            }
+        }
         return reactiveOperations.execute(preparedQuery);
     }
 }
